refactor: Fix ApiSuccess types and rename ApiError to match API docs
